Ignores UnicodeDecodeError and purchaseimport exceptions
